Back-End/filter-service/requirements.txt
fastapi
uvicorn
pillow
numpy
python-dotenv
python-multipart
openai
requests
transformers
torch